OK. I think I figured out the number of trajs thing.
When we say 10 or 50 in the yaml files, we actually get a few more, I think due to some quirk with multiprocessing. it shouldn't be a big deal. Just take the first X out of the values if we're interested in subsampling the trajectories for whatever reason. I subsampled at least 50 for all the pipelines (except humanoid which needed more). Now let's run ALL of the step 1 (training) for all four categories. I'm also rerunning the ones for the classical ones. Note that the paper also says they did 7 runs for those, so that makes sense. And whew, I feel like I understand this code now.